+++ MOP/src/FSM/FsmManager.cs
@@ -228,5 +228,17 @@ namespace MOP.FSM
                 return playerComputer.Value;
             }
         }
+
+        static FsmBool shadowsHouse;
+        public static bool ShadowsHouse
+        { 
+            get
+            {
+                if (shadowsHouse == null)
+                    shadowsHouse = GameObject.Find("Systems").transform.Find("Options").GetPlayMakerFSM("GFX").FsmVariables.GetFsmBool("ShadowsHouse");
+
+                return shadowsHouse.Value;
+            }
+        }
     }
 }